requirements.txt
streamlit
scikit-learn
polars
openpyxl
numpy
plotly
Pillow
xgboost
